requirements.txt
Django==2.1
django-crispy-forms==1.6.0
django-filter==2.1.0
django-markdown-deux==1.0.5
django-pagedown==0.1.1
django-rest-auth==0.9.5
djangorestframework==3.9.2
Markdown==3.1
markdown2==2.3.1
mysqlclient==1.4.2.post1
Pillow==3.1.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
